wip: rename from `core/commander` to `core/cli`
